+++ src/components/connectionInfoCard.tsx
@@ -6,35 +6,35 @@ import webuiVersion from "../libs/version";
 import { Wifi, CloseWifi } from "@icon-park/react";
 
 interface Props {
-  client: { current: Client | null };
+  client: Client | null;
 }
 export function ConnectionInfoCard(props: Props) {
   const update = useForceUpdate();
-  console.log(props.client.current?.data.pingStatus);
-  console.log(props.client.current?.data.memberIds);
+  // console.log(props.client?.data.pingStatus);
+  // console.log(props.client?.data.memberIds);
   useEffect(() => {
     const setListener = (m: Member) => {
       m.onPing.on(update);
     };
-    if (props.client.current && props.client.current.members().length > 0) {
-      setListener(props.client.current.members()[0]);
+    if (props.client && props.client.members().length > 0) {
+      setListener(props.client.members()[0]);
     } else {
-      props.client.current?.onMemberEntry.once(setListener);
+      props.client?.onMemberEntry.once(setListener);
       return () => {
-        props.client.current?.onMemberEntry.off(setListener);
+        props.client?.onMemberEntry.off(setListener);
       };
     }
   }, [props.client, update]);
   return (
     <Card title={`Connection Info`}>
-      <div>
+      <div className="w-full h-full overflow-auto">
         <p style={{ marginBottom: -4 }}>
           <span className="text-sm">Server:</span>
           <span className="font-mono pl-2">
-            {props.client.current?.serverName}
+            {props.client?.serverName}
           </span>
           <span className="font-mono pl-2">
-            {props.client.current?.serverVersion}
+            {props.client?.serverVersion}
           </span>
         </p>
         <p style={{ marginBottom: -4 }}>
@@ -49,13 +49,15 @@ export function ConnectionInfoCard(props: Props) {
           <span className="text-sm">Members:</span>
         </p>
         <ul>
-          {props.client.current?.members().map((m, i) => (
+          {props.client?.members().map((m, i) => (
             <li key={i} className="pl-4 flex items-center">
-              <span className="text-sm">{m.name}</span>
-              <span className="text-xs pl-1">({m.remoteAddr})</span>
-              <span className="text-xs font-mono pl-2">{m.libName}</span>
-              <span className="text-xs font-mono pl-1.5">{m.libVersion}</span>
-              <div className="pl-2">
+              <span className="text-sm">
+                {m.name}
+                {m.remoteAddr === "127.0.0.1" || (
+                  <span className="text-xs pl-0.5">({m.remoteAddr})</span>
+                )}
+              </span>
+              <div className="pl-1">
                 {m.pingStatus !== null ? (
                   <Wifi
                     fill={
@@ -74,6 +76,9 @@ export function ConnectionInfoCard(props: Props) {
                 {m.pingStatus !== null ? m.pingStatus : "?"}
               </span>
               <span className="text-sm pl-1">ms</span>
+              <span className="text-sm pl-1">:</span>
+              <span className="text-xs font-mono pl-1">{m.libName}</span>
+              <span className="text-xs font-mono pl-1.5">{m.libVersion}</span>
             </li>
           ))}
         </ul>