+++ src/libs/forceUpdate.ts
@@ -2,7 +2,7 @@ import { useState } from "react";
 
 //create your forceUpdate hook
 export function useForceUpdate() {
-  const [_, setValue] = useState(0); // integer state
+  const setValue = useState(0)[1]; // integer state
   return () => setValue((value) => value + 1); // update state to force render
   // A function that increment 👆🏻 the previous state like here
   // is better than directly setting `setValue(value + 1)`